Align ECCN history search with explorer
